refactored main.tsx
