+++ client/package.json
@@ -13,6 +13,7 @@
     "@apollo/client": "^3.7.14",
     "@headlessui/react": "^2.2.0",
     "@heroicons/react": "^2.2.0",
+    "@reduxjs/toolkit": "^2.6.0",
     "@tailwindcss/vite": "^4.0.1",
     "dayjs": "^1.11.13",
     "dotenv": "^16.4.6",
@@ -23,6 +24,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
+    "react-redux": "^9.2.0",
     "react-router-dom": "^6.28.0",
     "tailwindcss": "^4.0.1"
   },
