+++ client/src/App.tsx
@@ -1,47 +1,7 @@
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  createHttpLink,
-} from "@apollo/client";
-import { setContext } from "@apollo/client/link/context";
-import { SnackbarProvider } from "notistack";
-// import { Outlet } from "react-router-dom";
-
-import NavPanel from "./components/NavPanel.js";
-
-// Construct our main GraphQL API endpoint
-const httpLink = createHttpLink({
-  uri: "/graphql",
-});
-
-// Construct request middleware that will attach the JWT token to every request as an `authorization` header
-const authLink = setContext((_, { headers }) => {
-  // get the authentication token from local storage if it exists
-  const token = localStorage.getItem("id_token");
-  // return the headers to the context so httpLink can read them
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : "",
-    },
-  };
-});
-
-const client = new ApolloClient({
-  // Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-});
+import NavPanel from "./components/NavPanel";
 
 function App() {
-  return (
-    <ApolloProvider client={client}>
-      <SnackbarProvider>
-        <NavPanel />
-      </SnackbarProvider>
-    </ApolloProvider>
-  );
+  return <NavPanel />;
 }
 
 export default App;
